pyproject.toml
[project]
name = "oct24-bmlops-int-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.15.1",
    "annotated-types==0.7.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "databricks-sdk==0.46.0",
    "deepchecks>=0.19.1",
    "deprecated==1.2.18",
    "docker==7.1.0",
    "et-xmlfile==2.0.0",
    "evidently>=0.4.36",
    "fastapi>=0.116.1",
    "flask==3.1.0",
    "fonttools==4.56.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-auth==2.38.0",
    "graphene==3.4.3",
    "graphql-core==3.2.6",
    "graphql-relay==3.2.0",
    "greenlet==3.1.1",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "ipython>=9.0.2",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "joblib==1.4.2",
    "kiwisolver==1.4.8",
    "mako==1.3.9",
    "markdown==3.7",
    "markupsafe==3.0.2",
    "matplotlib==3.10.1",
    "mlflow==2.20.4",
    "mlflow-skinny==2.20.4",
    "numpy==2.2.3",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.31.0",
    "opentelemetry-sdk==1.31.0",
    "opentelemetry-semantic-conventions==0.52b0",
    "optuna>=4.2.1",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "protobuf==5.29.3",
    "pyarrow==19.0.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pyodbc==5.2.0",
    "pyparsing==3.2.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "pywin32==309",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rsa==4.9",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "seaborn>=0.13.2",
    "setuptools>=78.1.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sqlalchemy==2.0.38",
    "sqlparse==0.5.3",
    "threadpoolctl==3.6.0",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "waitress==3.0.2",
    "werkzeug==3.1.3",
    "wrapt==1.17.2",
    "xlsxwriter==3.2.2",
    "zipp==3.21.0",
]
